making progress on refactoring
